Create an abstract model class and reuse old code from model.py to define a new FeatureDetectionModel class
